refactor: クリーンアーキテクチャの違反箇所を修正
1. data 層の副作用を除去
- apiClient: 401 時の window.location.href リダイレクトを廃止し、
  UnauthorizedError を throw する方式に変更
- useFetcher (presentation 層) で UnauthorizedError を捕捉してリダイレクト
  → data 層を UI 副作用／SSR 依存から切り離し

2. useMissedDoses を DI 経由に統一
- ScheduleRepository に getMissedDoses / MissedDose を追加
- Client / Prisma の両実装にメソッドを追加
- scheduleApi.getMissedDoses() を追加
- useMissedDoses から apiClient 直 import を廃止し、DIContainer 経由に
  → 他の presentation hook と同じ DI パターンに揃えた
